refactor(message) : message와 관련된 클래스를 별도의 패키지로 생성
