+++ src/main/java/racingcar/message/UserMessage.java
@@ -1,4 +1,4 @@
-package racingcar.view;
+package racingcar.message;
 
 public class UserMessage {
     public static final String USER_CAR_NAME_INPUT_PROMPT = "경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).";
